Extract file size limit partial
